implements TodoFilter component; changes some logic in separate components
