+++ src/App.tsx
@@ -1,23 +1,34 @@
-/* eslint-disable jsx-a11y/control-has-associated-label */
-import React from 'react';
+import React, { useContext, useState } from 'react';
 import { TodoApp } from './components/TodoApp';
+import { Filter } from './types/Filter';
+import { TodosContext } from './components/TodosContext';
+import { TodoList } from './components/TodoList';
+import { TodoFilter } from './components/TodoFilter';
 
 export const App: React.FC = () => {
+  const [filter, setFilter] = useState<Filter>(Filter.All);
+  const { todos } = useContext(TodosContext);
+
+  const filtredTodo = todos.filter(({ completed }) => {
+    switch (filter) {
+      case Filter.Active:
+        return !completed;
+      case Filter.Completed:
+        return completed;
+      default:
+        return true;
+    }
+  }, [filter]);
+
   return (
     <div className="todoapp">
-
       <TodoApp />
-
-      {/* <section className="main">
-        <input
-          type="checkbox"
-          id="toggle-all"
-          className="toggle-all"
-          data-cy="toggleAll"
-        />
-        <label htmlFor="toggle-all">Mark all as complete</label>
-
-      </section> */}
+      {!!todos.length && (
+        <>
+          <TodoList items={filtredTodo} />
+          <TodoFilter filter={filter} onFilterChange={setFilter} />
+        </>
+      )}
     </div>
   );
 };
