+++ src/components/TodoList/TodoList.tsx
@@ -1,23 +1,48 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { Todo } from '../../types/Todo';
 import { TodoItem } from '../TodoItem';
+import { TodosContext } from '../TodosContext';
+import { ActionType } from '../../types/Action';
 
 type Props = {
   items: Todo[]
 };
 
 export const TodoList: React.FC<Props> = ({ items }) => {
+  const { dispatch } = useContext(TodosContext);
+
+  const allItemsCopmleted = items.every(({ completed }) => completed);
+
+  const handleToggleAll = () => {
+    dispatch({ type: ActionType.ChangeAllCompleted });
+  };
+
   return (
-    <ul className="todo-list" data-cy="todoList">
+    <section className="main">
+      <input
+        type="checkbox"
+        id="toggle-all"
+        className="toggle-all"
+        data-cy="toggleAll"
+        checked={allItemsCopmleted}
+        onChange={handleToggleAll}
+      />
+      <label htmlFor="toggle-all">Mark all as complete</label>
 
-      {items.map(item => (
-        <TodoItem
-          item={item}
-          key={item.id}
-        />
-      ))}
+      <ul className="todo-list" data-cy="todoList">
 
-      {/* <li className="completed">
+        {items.map(item => (
+          <TodoItem
+            item={item}
+            key={item.id}
+          />
+        ))}
+      </ul>
+    </section>
+  );
+};
+
+/* <li className="completed">
         <div className="view">
           <input type="checkbox" className="toggle" id="toggle-completed" />
           <label htmlFor="toggle-completed">qwertyuio</label>
@@ -57,7 +82,4 @@ export const TodoList: React.FC<Props> = ({ items }) => {
           />
         </div>
         <input type="text" className="edit" />
-      </li> */}
-    </ul>
-  );
-};
+      </li> */
